+++ aiccs-20230516/ChangeLog.md
@@ -1,3 +1,13 @@
+2024-08-22 Version: 1.0.4
+- Update API AddTask: add param FlashSmsTemplateId.
+- Update API AddTask: add param FlashSmsType.
+- Update API AddTask: update response param.
+- Update API EditTask: add param FlashSmsTemplateId.
+- Update API EditTask: add param FlashSmsType.
+- Update API EditTask: update response param.
+- Update API TaskList: update response param.
+
+
 2024-06-25 Version: 1.0.3
 - Update API TaskList: add param Status.
 - Update API TaskList: update response param.